requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
openai==1.54.3
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
setuptools==75.1.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
wheel==0.44.0
